Fix room space invite button for public spaces
